pre-commit.careful: use `git diff-index --raw -z` so as to correctly handle filenames with spaces and other odd characters
git-diff-index makes this rather difficult to get right. In its "unmunged"
output form, it changes both '\t' and '\n' delimiters to '\0', rather than
keeping them as distinct field/line delimiters.  >:-(

Also, in order for this to work, only bash will do, not any other /bin/sh, since only
bash seems to offer a way to use '\0' as a delimiter for the read builtin:

    http://transnum.blogspot.com/2008/11/bashs-read-built-in-supports-0-as.html?showComment=1399114879515#c545095514248554536

(thanks to @MareoRaft who brought this up in #12)